# 05_API-Specifications/openapi.yaml
openapi: 3.0.3
info:
  title: StarMitra API
  version: 1.0.0
  description: >-
    Canonical REST contract for StarMitra (ADR-007). Module-owned namespaces under
    /api/v1; DTO-only boundaries; RFC 9457 Problem Details; cursor/keyset
    pagination for streams, offset for bounded admin lists; database-enforced
    idempotency + Idempotency-Key header for retry-sensitive commands.
    Authentication per ADR-006: OTP + short-lived JWT (web httpOnly cookie,
    mobile Bearer) + opaque persisted refresh token.
servers:
  - url: /api/v1
...
paths:
...
  /admin/dashboard:
    get:
      tags: [Admin]
      operationId: adminDashboard
      summary: Operational dashboard view (Admin/SuperAdmin; derived, never business truth)
      responses: {'200': {description: Dashboard, content: {application/json: {schema: {type: object}}}}, default: {$ref: '#/components/responses/Problem'}}
  /admin/audit:
    get:
      tags: [Admin]
      operationId: adminAuditSearch
      summary: Platform AuditLog search view (kernel read-contract)
      parameters: [{name: module, in: query, schema: {type: string}}, {name: actorId, in: query, schema: {type: string, format: uuid}}, {name: correlationId, in: query, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Audit entries, content: {application/json: {schema: {type: object}}}}, default: {$ref: '#/components/responses/Problem'}}
  /admin/plans:
    get:
      tags: [Admin]
      operationId: adminListPlans
      summary: All plans incl. INACTIVE + configured entitlement values (M22-owned truth)
      responses: {'200': {description: Plans, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Plan'}}}}}, default: {$ref: '#/components/responses/Problem'}}
  /admin/plans/{planCode}:
    put:
      tags: [Admin]
      operationId: adminUpdatePlan
      summary: Update plan display/pricing/status (admin; M22 stays authoritative)
      parameters: [{name: planCode, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PlanUpdate'}}}}
      responses: {'200': {description: Updated plan, content: {application/json: {schema: {$ref: '#/components/schemas/Plan'}}}}, default: {$ref: '#/components/responses/Problem'}}
  /admin/plans/{planCode}/entitlements/{entitlementCode}:
    put:
      tags: [Admin]
      operationId: adminSetPlanEntitlement
      summary: Set one entitlement value on a plan (admin; interpreted per entitlement value_kind)
      parameters: [{name: planCode, in: path, required: true, schema: {type: string}}, {name: entitlementCode, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EntitlementSet'}}}}
      responses: {'200': {description: Updated plan, content: {application/json: {schema: {$ref: '#/components/schemas/Plan'}}}}, default: {$ref: '#/components/responses/Problem'}}

  # ---------------- M21 Social ----------------
...
  /public/plans:
    get:
      tags: [Pricing]
      operationId: listPlans
      summary: Public plan catalog — ACTIVE plans within their effective window
      security: []
      responses: {'200': {description: Plans, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Plan'}}}}}, default: {$ref: '#/components/responses/Problem'}}
  /me/plan:
    get:
      tags: [Pricing]
      operationId: myPlan
      summary: Caller's current plan + resolved entitlements (FREE by default)
      responses: {'200': {description: Current plan, content: {application/json: {schema: {$ref: '#/components/schemas/MyPlan'}}}}, default: {$ref: '#/components/responses/Problem'}}
    post:
      tags: [Pricing]
      operationId: selectPlan
      summary: Select a plan — FREE activates now; priced plans fail closed (402 PAYMENT_NOT_ENABLED)
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PlanSelect'}}}}
      responses: {'200': {description: Plan active, content: {application/json: {schema: {$ref: '#/components/schemas/MyPlan'}}}}, '402': {description: Payment integration not enabled}, default: {$ref: '#/components/responses/Problem'}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}              # mobile
    cookieAuth: {type: apiKey, in: cookie, name: sm_access}                  # web httpOnly+SameSite + CSRF token
...
  parameters:
...
    Cursor:        {name: cursor, in: query, schema: {type: string}, description: Opaque keyset cursor}
    Limit:         {name: limit, in: query, schema: {type: integer, default: 20, maximum: 100}}
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemDetails'}}}
  schemas:
    ProblemDetails:
      type: object
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        correlationId: {type: string, format: uuid}
        errors: {type: array, items: {type: object}}
      required: [type, title, status]
...
    Plan:           {type: object, properties: {id: {type: string, format: uuid}, code: {type: string}, displayName: {type: string}, description: {type: string}, planType: {type: string, enum: [SUBSCRIPTION, EVENT_PACKAGE]}, price: {type: number}, currency: {type: string}, billingPeriod: {type: string, enum: [NONE, MONTH, EVENT]}, status: {type: string}, sortOrder: {type: integer}, entitlements: {type: object, additionalProperties: {type: string}}}}
    PlanSelect:     {type: object, properties: {planCode: {type: string}}, required: [planCode]}
    PlanUpdate:     {type: object, properties: {displayName: {type: string}, description: {type: string}, price: {type: number, minimum: 0}, status: {type: string, enum: [ACTIVE, INACTIVE]}, sortOrder: {type: integer}, effectiveFrom: {type: string, format: date-time}, effectiveTo: {type: string, format: date-time}}}
    EntitlementSet: {type: object, properties: {value: {type: string}}, required: [value]}
    Entitlement:    {type: object, properties: {code: {type: string}, kind: {type: string, enum: [BOOLEAN, INTEGER, ENUM]}, value: {type: string}, limit: {type: integer, format: int64}, remaining: {type: integer, format: int64}}}
    MyPlan:         {type: object, properties: {planCode: {type: string}, displayName: {type: string}, subscriptionStatus: {type: string}, source: {type: string}, startsAt: {type: string, format: date-time}, endsAt: {type: string, format: date-time}, entitlements: {type: array, items: {$ref: '#/components/schemas/Entitlement'}}}}